feat: finalize invitation emails
