+++ backend/danswer/configs/app_configs.py
@@ -420,3 +420,5 @@ SECRET_JWT_KEY = os.environ.get("SECRET_JWT_KEY", "")
 
 DATA_PLANE_SECRET = os.environ.get("DATA_PLANE_SECRET", "")
 EXPECTED_API_KEY = os.environ.get("EXPECTED_API_KEY", "")
+
+ENABLE_EMAIL_INVITES = os.environ.get("ENABLE_EMAIL_INVITES", "").lower() == "true"
