+++ backend/danswer/server/utils.py
@@ -1,6 +1,16 @@
 import json
+import smtplib
 from datetime import datetime
+from email.mime.multipart import MIMEMultipart
+from email.mime.text import MIMEText
 from typing import Any
+
+from danswer.configs.app_configs import SMTP_PASS
+from danswer.configs.app_configs import SMTP_PORT
+from danswer.configs.app_configs import SMTP_SERVER
+from danswer.configs.app_configs import SMTP_USER
+from danswer.configs.app_configs import WEB_DOMAIN
+from danswer.db.models import User
 
 
 class DateTimeEncoder(json.JSONEncoder):
@@ -43,3 +53,28 @@ def mask_credential_dict(credential_dict: dict[str, Any]) -> dict[str, str]:
 
         masked_creds[key] = mask_string(val)
     return masked_creds
+
+
+def send_user_email_invite(user_email: str, current_user: User) -> None:
+    msg = MIMEMultipart()
+    msg["Subject"] = "Invitation to Join Danswer Workspace"
+    msg["To"] = user_email
+    msg["From"] = current_user.email
+
+    email_body = f"""
+Hello,
+
+You have been invited to join a workspace on Danswer.
+
+To join the workspace, please do so at the following link:
+{WEB_DOMAIN}/auth/login
+
+Best regards,
+The Danswer Team"""
+
+    msg.attach(MIMEText(email_body, "plain"))
+
+    with smtplib.SMTP(SMTP_SERVER, SMTP_PORT) as smtp_server:
+        smtp_server.starttls()
+        smtp_server.login(SMTP_USER, SMTP_PASS)
+        smtp_server.send_message(msg)
